+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Bug fix: remove lazy loading for backends, which is incompatible with precompilation.
+
 ## 0.2.27 (2025-04-06)
 * When `add`ing or `rm`ing a dependency, if resolving fails then CondaPkg.toml is now reverted.
 * Bug fixes.